moving uncertainty calculation, optimizing by only looping over half the correlation matrix
