+++ src/cabinetry/model_utils.py
@@ -1,6 +1,7 @@
 import logging
 from typing import List, Tuple
 
+import awkward1 as ak
 import numpy as np
 import pyhf
 
@@ -8,8 +9,8 @@ log = logging.getLogger(__name__)
 
 
 def get_parameter_names(model: pyhf.pdf.Model) -> List[str]:
-    """get the labels of all fit parameters, expanding vectors that act on
-    one bin per vector entry (gammas)
+    """Returns the labels of all fit parameters, expanding vectors that act on
+    one bin per vector entry (gammas).
 
     Args:
         model (pyhf.pdf.Model): a HistFactory-style model in ``pyhf`` format
@@ -76,3 +77,90 @@ def get_asimov_parameters(model: pyhf.pdf.Model) -> Tuple[np.ndarray, np.ndarray
                 pre_fit_unc += [0.0] * model.config.param_set(parameter).n_parameters
 
     return np.asarray(asimov_parameters), np.asarray(pre_fit_unc)
+
+
+def calculate_stdev(
+    model: pyhf.pdf.Model,
+    parameters: np.ndarray,
+    uncertainty: np.ndarray,
+    corr_mat: np.ndarray,
+) -> ak.highlevel.Array:
+    """Calculates the yield standard deviation of a model.
+
+    Args:
+        model (pyhf.pdf.Model): the model for which to calculate the standard
+            deviations for all bins
+        parameters (np.ndarray): central values of model parameters
+        uncertainty (np.ndarray): uncertainty of model parameters
+        corr_mat (np.ndarray): correlation matrix
+
+    Returns:
+        ak.highlevel.Array: array of channels, each channel
+        is an array of standard deviations per bin
+    """
+    # indices where to split to separate all bins into regions
+    # last index dropped since no extra split is needed after the last bin
+    region_split = [model.config.channel_nbins[ch] for ch in model.config.channels][:-1]
+
+    # the lists up_variations and down_variations will contain the model distributions
+    # with all parameters varied individually within uncertainties
+    # indices: variation, channel, bin
+    up_variations = []
+    down_variations = []
+
+    # calculate the model distribution for every parameter varied up and down
+    # within the respective uncertainties
+    for i_par in range(model.config.npars):
+        # best-fit parameter values, but one parameter varied within uncertainties
+        up_pars = parameters.copy()
+        up_pars[i_par] += uncertainty[i_par]
+        down_pars = parameters.copy()
+        down_pars[i_par] -= uncertainty[i_par]
+
+        # total model distribution with this parameter varied up
+        up_combined = model.expected_data(up_pars, include_auxdata=False)
+        up_yields = np.split(up_combined, region_split)
+        up_variations.append(up_yields)
+
+        # total model distribution with this parameter varied down
+        down_combined = model.expected_data(down_pars, include_auxdata=False)
+        down_yields = np.split(down_combined, region_split)
+        down_variations.append(down_yields)
+
+    # convert to awkward arrays for further processing
+    up_variations = ak.from_iter(up_variations)
+    down_variations = ak.from_iter(down_variations)
+
+    # total variance, indices are: channel, bins
+    total_variance_list = [
+        np.zeros(shape=(model.config.channel_nbins[ch])) for ch in model.config.channels
+    ]  # list of arrays, each array has as many entries as there are bins
+    total_variance = ak.from_iter(total_variance_list)
+
+    # loop over parameters to sum up total variance
+    # first do the diagonal of the correlation matrix
+    for i_par in range(model.config.npars):
+        symmetric_uncertainty = (up_variations[i_par] - down_variations[i_par]) / 2
+        total_variance = total_variance + symmetric_uncertainty ** 2
+
+    # continue with off-diagonal contributions if there are any
+    if np.count_nonzero(corr_mat - np.diag(np.diagonal(corr_mat))) > 0:
+        # off-diagonal contributions
+        # could optimize here: no need to loop over staterrors, since they
+        # contribute nothing (bin effects are orthogonal)
+        for i_par in range(model.config.npars):
+            for j_par in range(model.config.npars):
+                if j_par >= i_par:
+                    continue  # only loop over the half the matrix due to symmetry
+                corr = corr_mat[i_par, j_par]
+                if corr == 0:
+                    continue
+                sym_unc_i = (up_variations[i_par] - down_variations[i_par]) / 2
+                sym_unc_j = (up_variations[j_par] - down_variations[j_par]) / 2
+                # factor of two below is there since loop is only over half the matrix
+                total_variance = total_variance + 2 * (corr * sym_unc_i * sym_unc_j)
+
+    # convert to standard deviation
+    total_stdev = np.sqrt(total_variance)
+    log.debug(f"total stdev is {total_stdev}")
+    return total_stdev
